+++ src/redis/mutex.rs
@@ -1,7 +1,6 @@
 use crate::redis::Generic;
 use serde::de::DeserializeOwned;
 use serde::Serialize;
-use std::fmt::Display;
 use std::ops::{Deref, DerefMut};
 use thiserror::Error;
 
@@ -42,12 +41,12 @@ impl From<i8> for LockNum {
 /// 2. The timeout in seconds,
 /// 3. The value to store.
 const LOCK_SCRIPT: &str = r#"
-    local val = redis.call("get", ARGV[1] .. ":lock")
-    if val == false or val == ARGV[3] then
-        redis.call("setex", ARGV[1] .. ":lock", ARGV[2], ARGV[3])
-        return 1
-    end
-    return 0"#;
+local val = redis.call("get", ARGV[1] .. ":lock")
+if val == false or val == ARGV[3] then
+    redis.call("setex", ARGV[1] .. ":lock", ARGV[2], ARGV[3])
+    return 1
+end
+return 0"#;
 
 /// The drop script.
 /// It is used to drop a value in Redis, so that only the instance that locked it can drop it.
@@ -56,12 +55,12 @@ const LOCK_SCRIPT: &str = r#"
 /// 1. The key of the value to drop,
 /// 2. The value to check.
 const DROP_SCRIPT: &str = r#"
-    local current_lock = redis.call("get", ARGV[1] .. ":lock")
-    if current_lock == ARGV[2] then
-        redis.call("del", ARGV[1] .. ":lock")
-        return 1
-    end
-    return 0"#;
+local current_lock = redis.call("get", ARGV[1] .. ":lock")
+if current_lock == ARGV[2] then
+    redis.call("del", ARGV[1] .. ":lock")
+    return 1
+end
+return 0"#;
 
 /// The uuid script.
 /// It is used to generate a uuid for the lock.
@@ -104,6 +103,7 @@ end
 return nil"#;
 
 /// The RedisMutex struct.
+///
 /// It is used to lock a value in Redis, so that only one instance can access it at a time.
 /// You have to use RedisGeneric as the data type.
 /// It is a wrapper around the data type you want to store like the Mutex in std.
@@ -335,7 +335,7 @@ where
 
 impl<T> Deref for Guard<'_, T>
 where
-    T: DeserializeOwned + Serialize + Display,
+    T: DeserializeOwned + Serialize,
 {
     type Target = Generic<T>;
 
@@ -347,7 +347,7 @@ where
 
 impl<T> DerefMut for Guard<'_, T>
 where
-    T: DeserializeOwned + Serialize + Display,
+    T: DeserializeOwned + Serialize,
 {
     fn deref_mut(&mut self) -> &mut Self::Target {
         // Safety: The very existence of this Guard guarantees that we have exclusive access to the data.