+++ CHANGELOG.md
@@ -4,9 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
-## [Unreleased] - DESCRIPTION - YYYY-MM-DD
+## 0.2.2 - 2023-10-29
 
-- 
+- add ClockOrdered Type, which implements a counter based clock ordering
+- minor additions to documentation
 
 ## 0.2 - Lock and ADT - 2023-09-22